Align icons and text exactly with the pixel grid
Icon textures are (or can be) optimized to be displayed in a size of 48
pixels, which is a very common size to design for among icon designers.

Let's make sure we display the icons exactly as they were inteded to be shown
and use a fixed size of 48 pixels for them, aligning them perfectly with the
pixel grid by rounding the origin coordinates to physical pixels.

When used in combination with properly aligned icons, this eliminates all
the sub-pixel bluriness that can be seen on the edges of icons right now.

The text rendered for the function keys also becomes a lot less blurry
around the edges with this.